requirements.txt
Django==1.5.4
MySQL-python==1.2.4
South==0.8.2
django-tastypie==0.10.0
mimeparse==0.1.3
RestrictedPython==3.6.0
django_dajax==0.9.2
django-registration==1.0
